langgraph_studio/requirements.txt
# LangGraph Studio Requirements
# LangGraph Studio CLI
langgraph-cli>=0.1.0

# Core LangGraph and LangChain dependencies
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
httpx>=0.24.0

# Search tool dependency (optional)
tavily-python>=0.3.0

# Utility packages
regex>=2023.0.0